Split logic from Main to WritePathDemo
